requirements/dev.txt
-r base.txt

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=19.0,<20.0

# Code quality
black>=23.0,<24.0
isort>=5.12,<6.0
flake8>=6.0,<7.0

# Development
django-debug-toolbar>=4.2,<5.0
django-extensions>=3.2,<4.0
ipython>=8.0,<9.0
